requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
decorator==4.4.2
exceptiongroup==1.1.0
gTTS==2.3.0
h11==0.14.0
idna==3.4
imageio==2.23.0
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.24.1
outcome==1.2.0
packaging==22.0
Pillow==9.4.0
praw==7.6.1
prawcore==2.3.0
proglog==0.1.10
pycparser==2.21
pypiwin32==223
PySocks==1.7.1
pytesseract==0.3.10
pyttsx3==2.90
pytube==12.1.2
pywin32==305
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
update-checker==0.18.0
urllib3==1.26.13
websocket-client==1.4.2
wsproto==1.2.0
